管理员部分

+++ server/src/main/java/skk/controller/DDController.java
@@ -4,7 +4,10 @@ package skk.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import skk.entity.DD;
+import skk.entity.User;
 import skk.repository.DDRepository;
+import skk.repository.UserRepository;
+import skk.util.FailedResponse;
 import skk.util.Response;
 import skk.util.SuccessResponse;
 
@@ -24,10 +27,43 @@ class DDRequestBody{
 public class DDController {
 
     @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
     private DDRepository ddRepository;
 
-    @RequestMapping("/addOrUpdate")
-    public Response createEntry(@RequestBody DDRequestBody req){
+    User getUserInfo(String token){
+        List<User> users = userRepository.findByToken(token);
+        if(users.size() == 1){
+            return users.get(0);
+        } else {
+            return null;
+        }
+    }
+
+    public @GetMapping(path = "/entry")
+    Response allEntry(@RequestHeader(name = "x-skk-token", required = false , defaultValue = "null") String token){
+
+        User user = getUserInfo(token);
+        if(user == null){
+            FailedResponse r = new FailedResponse("身份认证失效，请重新登录");
+            return r;
+        }
+
+        List<DD> list = ddRepository.findAll();
+        SuccessResponse r = new SuccessResponse(list);
+        return r;
+    }
+
+    public @PostMapping(path = "/entry")
+    Response createEntry(@RequestHeader(name = "x-skk-token", required = false , defaultValue = "null") String token,
+                                @RequestBody DDRequestBody req){
+
+        User user = getUserInfo(token);
+        if(user == null){
+            FailedResponse r = new FailedResponse("身份认证失效，请重新登录");
+            return r;
+        }
 
         DD newDD = new DD();
         newDD.id = req.id;
@@ -36,25 +72,22 @@ public class DDController {
         newDD.code = req.code;
         newDD.entryValue = req.entryValue;
         ddRepository.save(newDD);
-        List<DD> list = ddRepository.findAll();
-        SuccessResponse r = new SuccessResponse(list);
-        return r;
+
+        return new SuccessResponse("修改成功");
     }
 
-    @RequestMapping("/delete")
-    public Response deleteEntry(@RequestBody DDRequestBody req){
+    public @DeleteMapping(path = "/entry")
+    Response deleteEntry(@RequestHeader(name = "x-skk-token", required = false , defaultValue = "null") String token,
+                                @RequestBody DDRequestBody req){
+
+        User user = getUserInfo(token);
+        if(user == null){
+            FailedResponse r = new FailedResponse("身份认证失效，请重新登录");
+            return r;
+        }
 
         ddRepository.deleteById(req.id);
-        List<DD> list = ddRepository.findAll();
-        SuccessResponse r = new SuccessResponse(list);
-        return r;
-    }
 
-    @RequestMapping("/all")
-    public Response showAll(){
-
-        List<DD> list = ddRepository.findAll();
-        SuccessResponse r = new SuccessResponse(list);
-        return r;
+        return new SuccessResponse("删除成功");
     }
 }
